+++ src/api/mod.rs
@@ -1,0 +1,8 @@
+pub mod middleware;
+pub mod routes;
+pub mod token_bucket;
+
+// Re-export public types and functions
+pub use middleware::log_request_errors;
+pub use routes::{create_claim, serve_video, upload_mp4_raw, waitlist};
+pub use token_bucket::TokenBucket;
