+++ src/lib.rs
@@ -1,17 +1,13 @@
 #![allow(incomplete_features)]
 #![feature(atomic_try_update, lazy_type_alias)]
 
+pub mod api;
 pub mod app_state;
 pub mod claim;
-pub mod claim_bucket;
-pub mod claim_middleware;
 pub mod config;
 pub mod job;
-pub mod middleware;
 pub mod opendal;
-pub mod routes;
 pub mod stream_map;
-pub mod token_bucket;
 
 use axum::{
     Router,
@@ -28,19 +24,19 @@ use tracing::info;
 //
 // Re-export
 //
+pub use api::{
+    TokenBucket, create_claim, log_request_errors, serve_video, upload_mp4_raw, waitlist,
+};
 pub use app_state::AppState;
 pub use claim::{
-    ClaimManager, ClaimPayloadV1, CreateClaimRequest, CreateClaimResponse, HLS_SEGMENT_DURATION,
+    ClaimBucketManager, ClaimManager, ClaimPayloadV1, ClaimState, CreateClaimRequest,
+    CreateClaimResponse, HLS_SEGMENT_DURATION, claim_auth_middleware,
     validate_claim_time_and_resource,
 };
-pub use claim_bucket::ClaimBucketManager;
-pub use claim_middleware::{ClaimState, claim_auth_middleware};
 pub use config::Config;
 pub use job::{ConvertJob, Job, JobResult, UploadJob};
 pub use opendal::{StorageBackend, StorageConfig, StorageManager};
-pub use routes::{create_claim, serve_video, upload_mp4_raw, waitlist};
 pub use stream_map::StreamMap;
-pub use token_bucket::TokenBucket;
 
 pub const RESOLUTIONS: [(u32, u32); 3] = [(720, 1280), (540, 960), (480, 854)];
 pub const BANDWIDTHS: [u32; 3] = [2500000, 1500000, 1000000];
@@ -112,7 +108,7 @@ pub async fn run(config: Config) {
         .allow_headers(Any);
 
     // Create claim state for middleware
-    let claim_state = claim_middleware::ClaimState {
+    let claim_state = claim::ClaimState {
         claim_manager: state.claim_manager.clone(),
         bucket_manager: state.claim_bucket_manager.clone(),
     };
@@ -122,9 +118,9 @@ pub async fn run(config: Config) {
         .route("/videos/{*filename}", get(serve_video))
         .route_layer(axum::middleware::from_fn_with_state(
             claim_state,
-            claim_middleware::claim_auth_middleware,
+            claim::claim_auth_middleware,
         ))
-        .layer(axum::middleware::from_fn(middleware::log_request_errors))
+        .layer(axum::middleware::from_fn(api::log_request_errors))
         .layer(cors.clone())
         .layer(Extension(state.clone()));
 
@@ -133,7 +129,7 @@ pub async fn run(config: Config) {
         .route("/claims", post(create_claim))
         .route("/upload", post(upload_mp4_raw))
         .route("/waitlist", get(waitlist))
-        .layer(axum::middleware::from_fn(middleware::log_request_errors))
+        .layer(axum::middleware::from_fn(api::log_request_errors))
         .layer(cors)
         .layer(Extension(state));
